Moved UI bootstrapping to TestAssembly. Enables mutiple test files.
